pyproject.toml
[tool.poetry]
name = "operator-cli"
version = "0.1.0"
description = "StakeWise Operator CLI is used to generate and manage ETH2 validator keys."
authors = ["Dmitri Tsumak <tsumak.dmitri@gmail.com>"]
license = "AGPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
python-decouple = "==3.6"
web3 = "==5.28.0"
py-ecc = "==5.2.0"
staking-deposit = {git = "https://github.com/ethereum/eth2.0-deposit-cli", rev = "v2.0.0"}
hvac = {version = "==0.11.2", extras = ["parser"]}
gql = {version = "==3.0.0", allow-prereleases = true, extras = ["aiohttp"]}
backoff = "==1.11.1"
aiohttp = {version = "==3.8.1", extras = ["speedups"]}
ssz = "==0.2.4"
click = "==8.0.3"
requests-toolbelt = "==0.9.1"
colorama = "==0.4.4"

[tool.poetry.dev-dependencies]
mypy = "==0.931"
black = "==22.1.0"
flake8 = "==4.0.1"
isort = "==5.10.1"
flake8-black = "==0.2.4"
flake8-bugbear = "==22.1.11"
pre-commit = "==2.17.0"
pyinstaller = "==4.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
